Working snapshot with writing to file with io_uring
some issues with timing that might be contributed
to the wrong linux headers used by conan and my kernel.
conan is using 5.13 and my kernel is 5.19.
Maybe this mismatch is the issue.

With sleep, I can have the io_uring completed
but it's not ideal.

Seeing:
onsume 2
result crc 0x5160
result crc 0x433